refactor(landing): reorganize all sections into subdirectory
