+++ ourportfolios/pages/landing/components/__init__.py
@@ -5,6 +5,7 @@ from .shiny_text import shiny_text, ShinyText
 from .magic_bento import magic_bento, magic_bento_card, MagicBentoCard
 from .card_swap import card_swap, card, CardSwapComponent, CardComponent
 from .scroll_reveal import scroll_reveal, ScrollRevealComponent
+from .badge_button import badge_button
 
 __all__ = [
     "plasma",
@@ -20,4 +21,5 @@ __all__ = [
     "CardComponent",
     "scroll_reveal",
     "ScrollRevealComponent",
+    "badge_button",
 ]
