+++ ourportfolios/pages/landing/sections/__init__.py
@@ -1,0 +1,15 @@
+"""Sections for the landing page."""
+
+from .hero_section import hero_section
+from .showcase_section import showcase_section
+from .bento_section import bento_section
+from .cta_section import cta_section
+from .footer import footer
+
+__all__ = [
+    "hero_section",
+    "showcase_section",
+    "bento_section",
+    "cta_section",
+    "footer",
+]
