+++ ourportfolios/pages/landing/sections/showcase_section.py
@@ -1,0 +1,384 @@
+"""Showcase section component with CardSwap demo."""
+
+import reflex as rx
+
+from ..components import shiny_text, scroll_reveal, card_swap, card
+
+
+def showcase_section() -> rx.Component:
+    """Showcase section with CardSwap demo."""
+    return scroll_reveal(
+        rx.center(
+            rx.box(
+                rx.hstack(
+                    rx.vstack(
+                        shiny_text(
+                            text="Card stacks have never looked so good",
+                            speed=3,
+                            color="#ffffff",
+                            shine_color="#ffffff",
+                            spread=120,
+                            direction="left",
+                            yoyo=False,
+                            delay=0,
+                            font_size=["2rem", "2.5rem", "3rem"],
+                            font_weight="600",
+                            line_height="1.2",
+                            letter_spacing="-0.02em",
+                        ),
+                        rx.text(
+                            "Just look at it go!",
+                            font_size=["1rem", "1.125rem"],
+                            color="rgba(255, 255, 255, 0.4)",
+                            margin_top="2rem",
+                            font_weight="300",
+                            line_height="1.5",
+                        ),
+                        align="start",
+                        spacing="0",
+                        flex="1",
+                        max_width="30rem",
+                        min_width="22rem",
+                    ),
+                    rx.box(
+                        rx.box(
+                            card_swap(
+                                card(
+                                    rx.box(
+                                        rx.hstack(
+                                            rx.icon(
+                                                "file_sliders",
+                                                size=20,
+                                                color="rgba(255, 255, 255, 0.5)",
+                                            ),
+                                            rx.heading(
+                                                "Customizable",
+                                                size="4",
+                                                font_weight="500",
+                                                letter_spacing="0.05em",
+                                                text_transform="uppercase",
+                                            ),
+                                            spacing="3",
+                                            align="center",
+                                            padding="1.5rem",
+                                        ),
+                                        rx.box(
+                                            rx.center(
+                                                rx.vstack(
+                                                    rx.heading(
+                                                        "Fully Customizable",
+                                                        size="5",
+                                                        font_weight="600",
+                                                        letter_spacing="-0.02em",
+                                                        margin_bottom="0.75rem",
+                                                    ),
+                                                    rx.text(
+                                                        "Tailor every aspect to match your workflow.",
+                                                        font_size="0.875rem",
+                                                        color="rgba(255, 255, 255, 0.6)",
+                                                        line_height="1.6",
+                                                        text_align="center",
+                                                    ),
+                                                    align="center",
+                                                    spacing="0",
+                                                ),
+                                                width="100%",
+                                                height="100%",
+                                            ),
+                                            padding="2rem",
+                                        ),
+                                        width="100%",
+                                        height="100%",
+                                        background="transparent",
+                                        backdrop_filter="blur(20px)",
+                                        border="1px solid rgba(255, 255, 255, 0.08)",
+                                        border_radius="1.5rem",
+                                        display="flex",
+                                        flex_direction="column",
+                                    ),
+                                ),
+                                card(
+                                    rx.box(
+                                        rx.hstack(
+                                            rx.box(
+                                                width="1.25rem",
+                                                height="1.25rem",
+                                                border_radius="9999px",
+                                                background="rgba(255, 255, 255, 0.5)",
+                                            ),
+                                            rx.heading(
+                                                "Smooth",
+                                                size="4",
+                                                font_weight="500",
+                                                letter_spacing="0.05em",
+                                                text_transform="uppercase",
+                                            ),
+                                            spacing="3",
+                                            align="center",
+                                            padding="1.5rem",
+                                        ),
+                                        rx.box(
+                                            rx.hstack(
+                                                rx.vstack(
+                                                    rx.box(
+                                                        rx.box(
+                                                            rx.text(
+                                                                "01",
+                                                                font_size="2rem",
+                                                                font_weight="900",
+                                                                color="rgba(255, 255, 255, 0.1)",
+                                                                margin_bottom="0.5rem",
+                                                            ),
+                                                            rx.text(
+                                                                "Fluid Interface",
+                                                                font_size="0.875rem",
+                                                                font_weight="500",
+                                                                letter_spacing="-0.02em",
+                                                            ),
+                                                        ),
+                                                        height="7rem",
+                                                        background="rgba(255, 255, 255, 0.02)",
+                                                        border="1px solid rgba(255, 255, 255, 0.06)",
+                                                        border_radius="0.875rem",
+                                                        padding="1.25rem",
+                                                    ),
+                                                    rx.hstack(
+                                                        rx.box(
+                                                            rx.center(
+                                                                rx.box(
+                                                                    rx.box(
+                                                                        width="66.67%",
+                                                                        height="100%",
+                                                                        background="#7C3AED",
+                                                                        border_radius="9999px",
+                                                                        box_shadow="0 0 6px rgba(124, 58, 237, 0.5)",
+                                                                    ),
+                                                                    width="2rem",
+                                                                    height="0.175rem",
+                                                                    background="rgba(255, 255, 255, 0.1)",
+                                                                    border_radius="9999px",
+                                                                    position="relative",
+                                                                ),
+                                                                height="100%",
+                                                            ),
+                                                            flex="1",
+                                                            height="4rem",
+                                                            background="rgba(255, 255, 255, 0.02)",
+                                                            border="1px solid rgba(255, 255, 255, 0.05)",
+                                                            border_radius="0.875rem",
+                                                        ),
+                                                        rx.box(
+                                                            flex="1",
+                                                            height="4rem",
+                                                            background="rgba(255, 255, 255, 0.02)",
+                                                            border="1px solid rgba(255, 255, 255, 0.05)",
+                                                            border_radius="0.875rem",
+                                                        ),
+                                                        spacing="3",
+                                                        width="100%",
+                                                    ),
+                                                    spacing="3",
+                                                    flex="1",
+                                                ),
+                                                rx.vstack(
+                                                    rx.vstack(
+                                                        rx.box(
+                                                            height="0.25rem",
+                                                            width="50%",
+                                                            background="rgba(255, 255, 255, 0.1)",
+                                                            border_radius="9999px",
+                                                        ),
+                                                        rx.box(
+                                                            height="0.25rem",
+                                                            width="75%",
+                                                            background="rgba(255, 255, 255, 0.05)",
+                                                            border_radius="9999px",
+                                                        ),
+                                                        rx.box(
+                                                            height="0.25rem",
+                                                            width="66.67%",
+                                                            background="rgba(255, 255, 255, 0.05)",
+                                                            border_radius="9999px",
+                                                        ),
+                                                        spacing="3",
+                                                    ),
+                                                    rx.spacer(),
+                                                    rx.center(
+                                                        rx.icon(
+                                                            "coins",
+                                                            size=28,
+                                                            color="rgba(255, 255, 255, 0.1)",
+                                                        ),
+                                                        width="100%",
+                                                        aspect_ratio="1",
+                                                        border_radius="0.625rem",
+                                                        background="linear-gradient(135deg, rgba(124, 58, 237, 0.1) 0%, transparent 100%)",
+                                                        border="1px solid rgba(255, 255, 255, 0.05)",
+                                                    ),
+                                                    spacing="3",
+                                                    flex="1",
+                                                    background="rgba(255, 255, 255, 0.02)",
+                                                    border="1px solid rgba(255, 255, 255, 0.05)",
+                                                    border_radius="0.875rem",
+                                                    padding="1.25rem",
+                                                    max_width="10rem",
+                                                ),
+                                                spacing="3",
+                                                align="stretch",
+                                                width="100%",
+                                                height="100%",
+                                            ),
+                                            padding="1.25rem",
+                                        ),
+                                        width="100%",
+                                        height="100%",
+                                        background="transparent",
+                                        backdrop_filter="blur(20px)",
+                                        border="1px solid rgba(255, 255, 255, 0.08)",
+                                        border_radius="1.5rem",
+                                        display="flex",
+                                        flex_direction="column",
+                                    ),
+                                ),
+                                card(
+                                    rx.box(
+                                        rx.hstack(
+                                            rx.text(
+                                                "</> ",
+                                                font_size="1rem",
+                                                color="rgba(255, 255, 255, 0.5)",
+                                            ),
+                                            rx.heading(
+                                                "Reliable",
+                                                size="4",
+                                                font_weight="500",
+                                                letter_spacing="0.05em",
+                                                text_transform="uppercase",
+                                            ),
+                                            spacing="3",
+                                            align="center",
+                                            padding="1.5rem",
+                                        ),
+                                        rx.box(
+                                            rx.center(
+                                                rx.vstack(
+                                                    rx.heading(
+                                                        "Reliable Performance",
+                                                        size="5",
+                                                        font_weight="600",
+                                                        letter_spacing="-0.02em",
+                                                        margin_bottom="0.75rem",
+                                                    ),
+                                                    rx.text(
+                                                        "99.9% uptime with enterprise-grade infrastructure.",
+                                                        font_size="0.8rem",
+                                                        color="rgba(255, 255, 255, 0.6)",
+                                                        line_height="1.6",
+                                                        text_align="center",
+                                                        margin_bottom="1.5rem",
+                                                    ),
+                                                    rx.hstack(
+                                                        rx.vstack(
+                                                            rx.text(
+                                                                "99.9%",
+                                                                font_size="1.75rem",
+                                                                font_weight="700",
+                                                                color="#7C3AED",
+                                                            ),
+                                                            rx.text(
+                                                                "Uptime",
+                                                                font_size="0.7rem",
+                                                                color="rgba(255, 255, 255, 0.4)",
+                                                            ),
+                                                            spacing="1",
+                                                            align="center",
+                                                        ),
+                                                        rx.vstack(
+                                                            rx.text(
+                                                                "<50ms",
+                                                                font_size="1.75rem",
+                                                                font_weight="700",
+                                                                color="#7C3AED",
+                                                            ),
+                                                            rx.text(
+                                                                "Latency",
+                                                                font_size="0.7rem",
+                                                                color="rgba(255, 255, 255, 0.4)",
+                                                            ),
+                                                            spacing="1",
+                                                            align="center",
+                                                        ),
+                                                        rx.vstack(
+                                                            rx.text(
+                                                                "24/7",
+                                                                font_size="1.75rem",
+                                                                font_weight="700",
+                                                                color="#7C3AED",
+                                                            ),
+                                                            rx.text(
+                                                                "Support",
+                                                                font_size="0.7rem",
+                                                                color="rgba(255, 255, 255, 0.4)",
+                                                            ),
+                                                            spacing="1",
+                                                            align="center",
+                                                        ),
+                                                        spacing="5",
+                                                    ),
+                                                    align="center",
+                                                    spacing="0",
+                                                ),
+                                                width="100%",
+                                                height="100%",
+                                            ),
+                                            padding="1.75rem",
+                                        ),
+                                        width="100%",
+                                        height="100%",
+                                        background="transparent",
+                                        backdrop_filter="blur(20px)",
+                                        border="1px solid rgba(255, 255, 255, 0.08)",
+                                        border_radius="1.5rem",
+                                        display="flex",
+                                        flex_direction="column",
+                                    ),
+                                ),
+                                width=550,
+                                height=400,
+                                card_distance=40,
+                                vertical_distance=45,
+                                delay=4000,
+                                pause_on_hover=True,
+                                skew_amount=0,
+                                easing="elastic",
+                            ),
+                            style={"transformStyle": "preserve-3d"},
+                        ),
+                        width=["100%", "100%", "650px"],
+                        max_width="650px",
+                        min_height="600px",
+                        display="flex",
+                        justify_content="center",
+                        align_items="flex-start",
+                        padding_top="2rem",
+                        margin_left=["0", "0", "3rem"],
+                    ),
+                    spacing="0",
+                    align="center",
+                    width="100%",
+                    max_width="1400px",
+                    padding_x=["1.5rem", "2rem", "4rem"],
+                    justify="center",
+                    gap=["4rem", "5rem", "6rem"],
+                    flex_direction=["column", "column", "row"],
+                ),
+                width="100%",
+                display="flex",
+                justify_content="center",
+                overflow="hidden",
+            ),
+            width="100%",
+            padding_y="10rem",
+            margin_top="6rem",
+        ),
+    )
